requirements.txt
python-binance
pynum
pandas
requests
fastapi
uvicorn
python-dotenv
